Move default root app fixture to operator to break test import cycle.
testutil importing config/providerpkg created a cycle through services/apps
tests, which failed the Go proto split dependency check in CI.
