+++ gestaltd/internal/operator/default_root_app_fixture.go
@@ -1,4 +1,4 @@
-package testutil
+package operator
 
 import (
 	"os"
