+++ gestaltd/internal/operator/testmain_test.go
@@ -178,7 +178,7 @@ func writeDefaultProvidersDir(baseDir, binaryPath string) (string, error) {
 	if err := writeProviderReleaseMetadata(dir, manifest); err != nil {
 		return "", err
 	}
-	if err := testutil.WriteDefaultRootAppFixture(providersDir); err != nil {
+	if err := WriteDefaultRootAppFixture(providersDir); err != nil {
 		return "", err
 	}
 	return providersDir, nil
